move slice and thin functions to utils module
