+++ src/dmd_era5/era5_download/__init__.py
@@ -2,14 +2,10 @@ from dmd_era5.era5_download.create_mock_era5 import create_mock_era5
 from dmd_era5.era5_download.era5_download import (
     config_parser,
     download_era5_data,
-    slice_era5_dataset,
-    thin_era5_dataset,
 )
 
 __all__ = [
     "config_parser",
     "download_era5_data",
     "create_mock_era5",
-    "slice_era5_dataset",
-    "thin_era5_dataset",
 ]
