+++ tests/era5_download/test_era5_download.py
@@ -11,9 +11,8 @@ from dmd_era5.era5_download import (
     config_parser,
     create_mock_era5,
     download_era5_data,
-    slice_era5_dataset,
-    thin_era5_dataset,
 )
+from dmd_era5.utils import slice_era5_dataset, thin_era5_dataset
 
 
 @pytest.fixture
